requirements.txt
click==8.3.1
dnspython==2.8.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
psycopg2-binary==2.9.11
pymongo==4.15.4
smmap==5.0.2
uvicorn==0.38.0
